Final Commit
Using Network Service, here I am reading the data which is of generic type and then parsing the JSON response by using the convertJson function which is also of generic type. After fetching the data, I am loading it in the UITableView. I have written few test cases on Network Service and Models.